Move hits component to instantsearch
