+++ src/components/instantsearch/hits.js
